bo0: wire psp second PlaySfxPositional copy as C in mixed TU (parametrized volume)
