+++ src/st/e_misc.h
@@ -958,31 +958,4 @@ void ClutLerp(RECT* rect, u16 palIdxA, u16 palIdxB, s32 steps, u16 offset) {
     }
 }
 
-void PlaySfxPositional(s16 sfxId) {
-    s32 posX, posY;
-    s16 sfxPan;
-    s16 sfxVol;
-
-    posX = g_CurrentEntity->posX.i.hi - 128;
-    sfxPan = (abs(posX) - 32) >> 5;
-    if (sfxPan > 8) {
-        sfxPan = 8;
-    } else if (sfxPan < 0) {
-        sfxPan = 0;
-    }
-    if (posX < 0) {
-        sfxPan = -sfxPan;
-    }
-    sfxVol = abs(posX) - 96;
-    posY = abs(g_CurrentEntity->posY.i.hi - 128) - 112;
-    if (posY > 0) {
-        sfxVol += posY;
-    }
-    if (sfxVol < 0) {
-        sfxVol = 0;
-    }
-    sfxVol = 127 - (sfxVol >> 1);
-    if (sfxVol > 0) {
-        g_api.PlaySfxVolPan(sfxId, sfxVol, sfxPan);
-    }
-}
+#include "e_play_sfx_positional.h"
