+++ src/st/e_play_sfx_positional.h
@@ -1,0 +1,34 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+#ifndef SFX_POSITIONAL_MAX_VOL
+#define SFX_POSITIONAL_MAX_VOL 127
+#endif
+
+void PlaySfxPositional(s16 sfxId) {
+    s32 posX, posY;
+    s16 sfxPan;
+    s16 sfxVol;
+
+    posX = g_CurrentEntity->posX.i.hi - 128;
+    sfxPan = (abs(posX) - 32) >> 5;
+    if (sfxPan > 8) {
+        sfxPan = 8;
+    } else if (sfxPan < 0) {
+        sfxPan = 0;
+    }
+    if (posX < 0) {
+        sfxPan = -sfxPan;
+    }
+    sfxVol = abs(posX) - 96;
+    posY = abs(g_CurrentEntity->posY.i.hi - 128) - 112;
+    if (posY > 0) {
+        sfxVol += posY;
+    }
+    if (sfxVol < 0) {
+        sfxVol = 0;
+    }
+    sfxVol = SFX_POSITIONAL_MAX_VOL - (sfxVol >> 1);
+    if (sfxVol > 0) {
+        g_api.PlaySfxVolPan(sfxId, sfxVol, sfxPan);
+    }
+}
